meta: move server code to ui::server
Now `cli/ui.rs` contains code relevant to the ui CLI command, and the code
for the axum server itself is in `cli/ui/server.rs`. Some cross-referencing
between them still happens.

Also, switch from using `kill` from the `nix` crate on Windows to using
`taskkill`.
